Refactor pmap implementation into pxla.py.
This is setup for being able to recursively call pmap for multihost computations.
